zelus-website-api/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
psycopg2-binary
bcrypt
pydantic[email]
slowapi
requests
python-dotenv
stripe
httpx
